Bullpen Information

Information on venirepersons who were not disqualified or exempted.

Usage

bullpen

Format

A tibble with 3128 rows and 62 variables:

juror_id

Unique juror identifier.

venireperson

Uique juror identifier per trial.

caused

Venireperson was challenged and struck for cause.

state

State used peremptory challenge against venireperson.

defense

Defense used peremptory challenge against venireperson.

juror_not_struck

Juror status.

ammend_5

Caused for inability to accept defendant's 5th Amendment rights.

judgement

Caused for inability to stand in judgement due to religious or moral beliefs.

inn

Caused for inability to presume a defendant's innocence.

bias_no_fair

Venireperson was challenged and struck for bias.

bias_for_state

Venireperson was challenged and struck for bias for the state.

bias_a_g_state

Venireperson was challenged and struck for bias against the state.

bias_for_def

Venireperson was challenged and struck for bias for the defense.

bias_a_g_def

Venireperson was challenged and struck for bias against the defense.

bias_a_g_police

Venireperson was challenged and struck for bias against the police.

bias_for_police

Venireperson was challenged and struck for bias for the police.

cant_follow_law

Venireperson was challenged and struck for inability to follow the law.

attn

Venireperson was challenged and struck for inability to pay attention.

bop_100_percent

Venireperson was challenged and struck for requiring a higher burden of proof than beyond a reasonable doubt.

phys

Venireperson was challenged and struck for requiring physical evidence.

agmt

Venireperson was challenged and struck by agreement between the State and Defense.

vict

Venireperson was challenged and struck for prior exerience as a victim.

eye_witness

Venireperson was challenged and struck for being unable to accept only one witness.

pun

Venireperson was challenged and struck for being unable to accept the full range of possible punishment.

pun_min

Venireperson was challenged and struck for being unable to accept the minimum punishment.

pun_max

Venireperson was challenged and struck for being unable to accept the maximum punishment.

esl

Venireperson was unable to read and write English.

covid

Venireperson was uncomfortable serving on a jury during COVID-19.

med

Venireperson had medical issues that made them unfit for jury service.

prior_conviction

Venireperson had a prior conviction that made them ineligible for jury service.

no_show

Venireperson checked in to the courthouse, but did not appear at the courtroom.

work_exempt

Venireperson was excused for a work exemption.

child_exempt

Venireperson was excused for a childcare exemption.

age_exempt

Venireperson was excused for an age exemption.

jur_age

Venireperson's age.

jur_race

Venireperson's race.

jur_gend

Venireperson's gender.

zip_code

Venireperso's zip code.

four

4 digits after zip code.

defendants_name

Anonymized name for each defendant.

defendants_race

Defendant's race.

defendants_gender

Defendant's gender.

cause_number_f_xx_xxxxx

Anonymized trial number.

charge

The charge against the defendant.

verdict

The trial's verdict.

punishment_number_years

The term of the punishment..

probation

Defendant was put on probation.

notes

Notes made by the Judge during jury selection.

age_group

Venireperson's age binned into a group.

peremptory_strike

Summary variable for State and Defense.

jur_caused

Summary variable for dismissed for cause.

jur_struck

Summary variable for struck.

jur_exempt

Summary variable for exempt from service.

jur_disqualified

Summary variable for disqualified.

jur_biased

Summary variable for bias.

jur_pun

Summary variable for punishment.

city

City associated with juror's zip code.

i-30

Location of zip code in relation to Interstate 30.

po_box

Zip code is associated with a P.O. Box.

jur_race_w

Same as jur_race, but with "White" as the reference level.